Item8640:  Add utility to copy extension into the live system.
I don't have a testcase for this, but it seems to work even if the /tmp location is on a different file system.
The "move" for directories would fail on Linux when moving across file system so this now creates the
directory before moving the files.  It also seems to deal with readonly files, and files existing where a
directory will go.

Needs testing on other platforms!



git-svn-id: http://svn.foswiki.org/trunk/core@6639 0b4bb1d4-4e5a-0410-9cc4-b2b747904278