Item8640:  Create a generic installer for extensions.
This was accidental, but with the refactoring of extender.pl, it is now very 
easy to create a generic installer for packages. It is a 2 line change from a
template_installer package.  "shift" to recover the package name from the
command line and pass it to the call to extender. 


git-svn-id: http://svn.foswiki.org/trunk@7083 0b4bb1d4-4e5a-0410-9cc4-b2b747904278