Revert Foswikirev:7551 "Item9062: remove Compatibility plugin - its optional for most of us"
This plugin should simply be disabled, but not removed. Configure would be
smart enough to guess the settings in order to enable it, but one shouldn't
rely on that.

Also added some code in FoswikiTestCase which will yell when trying to enable a
plugin which doesn't have a $Foswiki::cfg{Plugins}{$plugin}{Module}, and guess
the module name to be 'Foswiki::Plugins::' . $plugin;

git-svn-id: http://svn.foswiki.org/trunk@8198 0b4bb1d4-4e5a-0410-9cc4-b2b747904278