Item9317: Alias body to head zone wasn't good enough
Instead, we keep both zones, but if {OptimizePageLayout} not set, then
do special things to check the opposite zone when trying to walk the
requirements and when building the list of ZoneIDs to iterate over (for
applying the format param) we actually append the body records to the head
list before visiting

Added some stuff to the tests to stress this a little more. Also, test
that multiple ADDTOZONE calls to the same id will in fact overwrite instead
of add.

Changed the way we report missing requirements; they are pushed to an array
for any ZoneID whose requirements are being visited

git-svn-id: http://svn.foswiki.org/trunk@8309 0b4bb1d4-4e5a-0410-9cc4-b2b747904278