Item10413: Downgrade Locale::Maketext::Lexicon
Version 0.86 requires version 1.17 of Locale::Maketext.  Version 0.79 of
Locale::Maketext::Lexicon still has the perl 5.12 fixes but does not
require an updated Maketext.  This is probably safer to ship.

Also remove Maketext.pm from manifest and CPAN lib.  And add
Locale::Maketext to the dependencies so that it will show up in
configure for debugging purposes.

git-svn-id: http://svn.foswiki.org/branches/Release01x01/core@10900 0b4bb1d4-4e5a-0410-9cc4-b2b747904278