Item1562: redirect denied ok button wrong
The redirect_denied error uses the generic oops message and the ok
button goes to the WebHome topic instead of the newly renamed topic.

Use a new redirectdenied template, and alter the current session web and
topic names so that the OK button will return to the newly renamed topic
should the redirect fail.

git-svn-id: http://svn.foswiki.org/branches/Release01x01/core@11029 0b4bb1d4-4e5a-0410-9cc4-b2b747904278