Item10927: Fix destruction of meta when you update attach props
Foswiki::Meta->new() is ... not cool. Use Foswiki::Meta::readTopic()
instead.

The only reason this appears to work here, is because the checkAccess
must load the topic fully in order to process ACLs. But if you're an
admin user, then the $topicObject is never loaded (doesn't need to
check ACLs, so the topic isn't loaded).

So any admin user trying to update properties on an attachment, will
delete all meta on that topic (we end up saving a $topicObject that
was never loaded).

git-svn-id: http://svn.foswiki.org/branches/Release01x01/core@12082 0b4bb1d4-4e5a-0410-9cc4-b2b747904278