Item9164: Fix crypt-md5 encoding for windows.
This most likely also fixes OSX, but needs someone to comment out the
exception and run the unit tests.  Issue is that the crypt extension for
md5 encoding is a Linux glibc extension.

This commit also adds limited auto-detect for plain text passwords.
However a 13 character plain text password will be mis-identified as
crypt.

git-svn-id: http://svn.foswiki.org/trunk/core@12208 0b4bb1d4-4e5a-0410-9cc4-b2b747904278