Item11120: Add javascript disabled error
The validate confirmation dialog was fixed to report that confirmation
was not possible due to disabled javascript.  However that's rather late
if a user has done a large number of edits.  Add a warning at the top of
the edit screen.   Pulled the validate message into a common message
template so it can be reused in multiple places.

git-svn-id: http://svn.foswiki.org/trunk/core@12498 0b4bb1d4-4e5a-0410-9cc4-b2b747904278