Item11196: Better feedback from UserMapping
Sync changes from trunk

The TopicUserMapping and UserMapping interfaces don't use maketext to
translate any user messages. And they in several places return 0 for
failures with no indication of the errors.

Note that this changes the API for the mappers. The mapper should throw
Error::Simple when encountering issues rather than leaving the user
guessing.

git-svn-id: http://svn.foswiki.org/branches/Release01x01/core@13744 0b4bb1d4-4e5a-0410-9cc4-b2b747904278