Item11448: Don't mask other errors
The change to catch the redirect to do header cleanup might be masking
other Error::Simple thrown by the topic save.  So if the Error::Simple
is not the restauth redirect, propagate it on up.

Also add a checker to warn if GuestCanComment is inconsistent with the
protection of the rest and restauth scripts.

Finally add a warning to the {AuthScripts} checker.  If any *auth script
is not listed, we need to warn the admin.

git-svn-id: http://svn.foswiki.org/trunk/core@13746 0b4bb1d4-4e5a-0410-9cc4-b2b747904278