Item11612: Use IPC::Run in Foswiki::Net
It does a much better job of supporting capture of STDERR.

Still need to move Foswiki::Sandbox to IPC::Run, but that will be a lot
more complicated to test - as Sandbox is used in many places for lots of
stuff.   Foswiki::Net uses it to call sendmail,  so the platforms where
this will apply are pretty narrow in scope.

git-svn-id: http://svn.foswiki.org/trunk/core@15543 0b4bb1d4-4e5a-0410-9cc4-b2b747904278