Item12257:  Still didn't get the checker right.
If a Warning file is set,  and the user chose PlainFile,  then it's an
error because the system will silently override back to compatibility in
Foswiki.pm.   But if it's any other logger except Compatibility, just
note that the setting is deprecated.

If warning file is NOT set,  and user chose Compatibility, then it's a
warning because the logger will assume a default location.

And also, don't modify the configuration settings.  Foswiki.pm will do
the fallback if needed.  And modifying the setting in the Implementation
checker causes the individual field checkers to mask their errors.

Also, simplify the checkers a bit.  The error is reported in multiple
places with redundant information.

git-svn-id: http://svn.foswiki.org/trunk/core@16081 0b4bb1d4-4e5a-0410-9cc4-b2b747904278