Item12464: Better instructions for reverting changes
Cut/paste for reverting a topic is complicated and fails to revert
metadata.  Document the Restore in More topic actions.

Also, note the limitations about removing sensitive information from
history.

Add the delRev and repRev options to CommandAndCGIScripts for
the edit script.  They were already documented on the save script.

Note that the "Put it back" link is now shown in the More topic actions
page and not at the bottom of the topic.

Note that when cleaning the trash, the TrashAttachments topic should be
recreated.

git-svn-id: http://svn.foswiki.org/trunk/core@16670 0b4bb1d4-4e5a-0410-9cc4-b2b747904278